Fixes script reading and encoding problem.
It seems that I've finally solved the old problem with encodings. And I
was on a wrong path. We don't need to convert any characters at all!
I just used FileReader to read scripts. And this class use a system
default encoding. For Windows, it's Windows-1251. And it was a bad idea!
I'm so ashamed that I couldn't find such silly mistake for a long time
and pushed pseudo-fixes that made the situation worse and worse...