Adds new plugins: `server` and `test_server`.
The former is responsible for listening for socket connections and
proxying all JSON requests to the message system.
The latter demonstrates the concept.

The extension sends messages to the Groovy plugin `test_server`, which
should do the rest logic.
The `test.py` script is just an executable Python script. Run it from
the terminal to test the `server` plugin.